Special:Nuke: Remove submit button at the top of the form
There's already another submit button at the bottom of the form.
Since this button at the top is very close to the select links
and it is easy to unintentionally click on the submit button
while attempting to click a select link, remove the button at
the top. A submit button at the bottom is also how we normally
have in other forms.

Bug: T145346
Change-Id: If0b21aea8ef8f1762bdafc98527de06f04d59703