Switch to recentchanges in pattern-only search
This forces the use of the `recentchanges` table in cases where an
actor target is not provided. It turns out that running queries on
`revision` is extremely inefficient when there's no `actor`, so
this second path has to be followed to ensure that the tool still
works, at the cost of reducing its time range.

This introduces a special query builder specifically for the Nuke
query, which splits apart functionality and abstracts the switching
between tables.

A message will be displayed when this specific behavior is being
followed (i.e. when $wgNukeMaxAge > $wgRCMaxAge).

Tests have been adjusted accordingly. A new trait has been split
off from existing tests to share similar utility functions that are
useful in other tests.

Bug: T380846
Change-Id: I708b46d634d695209f7aed1e90f221f2798b22eb