Introduce context system for form params
For every new filter added to Nuke, the method signature for each step
in the form grows larger. To alleviate this, this change introduces a
context system which will be used to centralize already-sanitized
request parameters and helper functions. This means adding more
filters or swapping out parts of the interface shouldn't cause a
significant refactor every time that requires changing signatures or
duplication of sanitization/validation steps.

This is lead up work to eventual Codex conversion, but also to
support the increasing number of filters being added to the prompt
form.

Change-Id: I3332586f673bd6b395a64f818f31519c9a8aa3f4