Add Codex UI renderer
Adds a Codex-based UI renderer, which renders the Nuke form using
Codex PHP.

Since we're expecting to scrap a lot of the current UI down the line
when converting to a fully-Codex setup, there's a lot of duplication
between the HTMLForm UI and the Codex UI. This should slowly be
removed later on as the new UI is designed.

Tests for this patch currently derive from the standard tests which
run against the HTMLForm version of the UI. This should be changed
down the line when the UI begins to deviate from the original.

The JavaScript configuration for Nuke has been bumped to support
the latest `wikimedia/client` style and Vue. Vue + Codex is used to
implement the namespace MultiselectLookup.

Bug: T153988
Change-Id: Ia45387802d89d16c32d99845f48f996845a54363