Show underlying status message in results
Currently, Nuke only indicates that a page has been deleted or not
deleted without showing the underlying reason to the error. This can
be rather cryptic to the user, and is in general bad UX practice.

This exposes the underlying status message returned by the deletion
to ensure that we show the reason why a deletion failed to the user.
This applies specifically to file deletions and not page deletions,
which use a job (and because of that, we can't get their status at
page render time).

Bug: T402507
Change-Id: Ief35d0337490e932359e836c93255a821e0af381